requirements.txt
numpy
nltk
pytorch
transformers
transformers[torch]
evaluate
datasets=2.15.0
sentencepiece